requirements.txt
binance-connector==1.1.1
black==21.6b0
Django==3.2.5
djangorestframework==3.12.4
python-decouple==3.4
pytz==2021.1
